Modified simulation so that parameters can be changed
Parameters can now change after a certain run in a simulation with multiple runs.

+++ src/genome-compare.cpp
@@ -189,7 +189,7 @@ float geneticDiversitySpecifiedPopulation(std::string species)
 
     if (species == "mouse") {
         // count limits the number of genomes of mice sampled for performance reasons.
-        unsigned numMice = p.population * p.miceRatio;
+        unsigned numMice = static_cast<uint16_t>(p.population * p.miceRatio);
 
         if (numMice < 2) {
             return 0.0;
@@ -211,7 +211,7 @@ float geneticDiversitySpecifiedPopulation(std::string species)
         return diversity;
     } else {
         // count limits the number of genomes of cats sampled for performance reasons.
-        unsigned numCats = p.population - p.population * p.miceRatio;
+        unsigned numCats = static_cast<uint16_t>(p.population - p.population * p.miceRatio);
 
         if (numCats < 2) {
             return 0.0;
